Move StdDevSamp into the new algebraic aggregate source files.
