+++ src/include/function/aggregate_function/list.hpp
@@ -1,1 +1,2 @@
 #include "function/aggregate_function/distributive.hpp"
+#include "function/aggregate_function/algebraic.hpp"
